Update to SDK version 6.0
This plug-in still works fine in Cura 4.0, but the view mode looks a bit weird since it has no extra button in the middle of the top bar. So the settings menu shifts towards the middle of the screen and then obscures the view a bit. It's slightly annoying but workable if you're debugging (because you probably don't need access to the settings if you're debugging scene-related stuff then).